refactor: colocate recorder project state type
