+++ src/lib/recorder/runtime.ts
@@ -4,6 +4,7 @@ import { type AudioView, createAudioView } from "../audio-view.ts";
 import { AudioBufferPlayback } from "./audio-buffer-playback.ts";
 import { CaptureInput } from "./capture-input.ts";
 import { RecorderMetronome } from "./metronome.ts";
+import type { RecorderProjectState } from "./project-persistence.ts";
 import { ActiveRecording } from "./recording.ts";
 import { AudioContextTransport } from "./transport.ts";
 
@@ -61,16 +62,6 @@ export interface RecorderRuntimeState {
   selectedChannel: number;
   latencyCompensation: number;
 }
-
-export type RecorderProjectState = Pick<
-  RecorderRuntimeState,
-  | "title"
-  | "tempo"
-  | "timeSignature"
-  | "audioTracks"
-  | "recordingTrack"
-  | "latencyCompensation"
->;
 
 const METRONOME_GAIN = 0.5;
